# feat(rust/catalyst-voting): Parallel iter usage for `fn generate_unit_vector_proof`