# Description

Updated `generate_unit_vector_proof` function with the parallel iterator usage.

Performance measurements on `Apple M1`
Old version:
```
vote protocol benchmark/voter proof generation
                        time:   [651.54 µs 651.66 µs 651.91 µs]
```

New version:
```
vote protocol benchmark/voter proof generation
                        time:   [454.23 µs 457.56 µs 464.87 µs]
```